requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1 